pyproject.toml
[project]
name = "spotify-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "pydantic-settings>=2.13.1",
    "spotipy>=2.26.0",
]

[tool.uv.workspace]
members = [
    "fast",
]
